Moving SliderConverter to Tactical Data.
