+++ org.mwc.debrief.lite/src/main/java/org/mwc/debrief/lite/menu/DebriefRibbonTimeController.java
@@ -68,10 +68,10 @@ import MWC.GUI.Tools.Swing.MyMetalToolBarUI.ToolbarOwner;
 import MWC.GUI.Undo.UndoBuffer;
 import MWC.GenericData.HiResDate;
 import MWC.GenericData.TimePeriod;
+import MWC.TacticalData.SliderConverter;
 import MWC.TacticalData.temporal.ControllablePeriod;
 import MWC.TacticalData.temporal.PlotOperations;
 import MWC.TacticalData.temporal.TimeManager;
-import junit.framework.TestCase;
 
 public class DebriefRibbonTimeController
 {
@@ -219,61 +219,6 @@ public class DebriefRibbonTimeController
       final Point p = component.getLocationOnScreen();
 
       menu.setLocation(p.x, p.y + component.getHeight());
-    }
-  }
-
-  /**
-   * utility class to handle converting between slider range and time values
-   *
-   * @author ian
-   *
-   */
-  private static class SliderConverter
-  {
-    private int range;
-    private long origin;
-    // have one minute steps
-    private final int step = 1000;
-
-    public int getCurrentAt(final long now)
-    {
-      return (int) Math.round((double) (now - origin) / step);
-    }
-
-    public int getEnd()
-    {
-      return range;
-    }
-
-    public int getStart()
-    {
-      return 0;
-    }
-
-    public long getTimeAt(final int position)
-    {
-      return origin + (position * step);
-    }
-
-    public void init(final long start, final long end)
-    {
-      origin = start;
-      range = (int) ((end - start) / step);
-    }
-  }
-
-  public static class SliderConverterTest extends TestCase
-  {
-    public void testConverter()
-    {
-      final SliderConverter test = new SliderConverter();
-      test.init(1240423198490L, 1240427422390L);
-
-      final int originalStep = 21;
-      final long originalTime = test.getTimeAt(originalStep);
-      final long roundedTime = originalTime / 1000L * 1000L;
-      final int newStep = test.getCurrentAt(roundedTime);
-      assertEquals("Rounding slider converter", originalStep, newStep);
     }
   }
