+++ org.mwc.debrief.test2/src/org/mwc/debrief/test/AllTests.java
@@ -147,7 +147,7 @@ import org.junit.runners.Suite;
     org.mwc.cmap.TimeController.recorders.CoordinateRecorder.CoordinateRecorderTest.class,
     org.mwc.debrief.core.ContextOperations.SmoothTrackJumps.TestMe.class,
     Debrief.ReaderWriter.BRT.BRTImporter.BRTImporterTest.class,
-    org.mwc.debrief.lite.menu.DebriefRibbonTimeController.SliderConverterTest.class
+    MWC.TacticalData.SliderConverter.SliderConverterTest.class
     })
 @RunWith(Suite.class)
 public class AllTests
